Major Code Update
Rewrote the Setup command to enable a Scoop feature/clean up code.
Modifed the Tap function to switch it to a batchExecute.
Changed the Play/Charge alignment system to be more robust.
Improved the Combo system and message for it.
Removed the Shared pile as it serves no purpose.
Added an end of turn cleanup function.
Deleted a number of extraneous functions left over from old templates.